feat: 중량 별 배송비의 기준 가격 변경 및 항공 배송비의 추가 #16081

+++ app/Abstracts/ProductAbstract.php
@@ -306,10 +306,10 @@ abstract class ProductAbstract
      * @func weightSave
      * @description '상품 중량 저장'
      * @param array $offerIds '제품 ID'
-     * @param int $weight '표준 중량'
+     * @param float $weight '표준 중량'
      * @return array
     */
-    abstract function weightSave(array $offerIds, int $weight): array;
+    abstract function weightSave(array $offerIds, float $weight): array;
 
     /**
      * @func noticeNameUpdate
